+++ support/installer/src/ComposerInstaller.php
@@ -18,9 +18,10 @@ class ComposerInstaller extends LibraryInstaller
 	public function supports($packageType)
 	{
 		return in_array($packageType, [
+			'heroku-sys-hhvm',
+			'heroku-sys-library',
 			'heroku-sys-php',
 			'heroku-sys-php-extension',
-			'heroku-sys-hhvm',
 			'heroku-sys-webserver',
 		]);
 	}